IOS_App(Calculator)_Testing_On_BrowserStack/requirements.txt
pytest
appium-python-client
python-dotenv
